requirements.txt
Flask==2.3.3
openai==1.3.7
PyPDF2==3.0.1
python-docx==0.8.11
numpy==1.24.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.29.2
huggingface-hub==0.14.1
uuid
python-dotenv==1.0.0
pandas==1.5.3
scikit-learn==1.3.0
Werkzeug==2.3.7
requests==2.31.0
nltk==3.8.1
spacy==3.6.1
regex==2023.8.8
